src/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "sharepoint-gpt-middleware"
version = "0.1.0"
description = "An Azure Web App in Python to crawl the content of SharePoint sites and perform semantic search using OpenAI Vector Stores."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Karsten Held" }]
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "cryptography==43.0.3",
  "azure-identity==1.24.0",
  "requests==2.32.5",
  "python-dotenv==1.1.1",
  "openai==1.101.0",
  "azure-keyvault-secrets==4.10.0",
  "aiohttp==3.12.15",
  "python-multipart==0.0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "httpx>=0.27",
  "ruff>=0.5"
]

# Single-module layout (app.py in this folder)
[tool.hatch.build.targets.wheel.force-include]
"app.py" = "app.py"

[tool.hatch.build.targets.sdist]
include = [
  "**",
  "README.md",
  "pyproject.toml"
]

[tool.ruff]
line-length = 200
